1.2

+++ src/rpcmining.cpp
@@ -69,7 +69,7 @@ Value getstakinginfo(const Array& params, bool fHelp)
     uint64_t nWeight = 0;
     pwalletMain->GetStakeWeight(nWeight);
 
-    uint64_t nNetworkWeight = GetPoSKernelPS();
+    double nNetworkWeight = GetPoSKernelPS();
     bool staking = nLastCoinStakeSearchInterval && nWeight;
     uint64_t nExpectedTime = staking ? (GetTargetSpacing(nBestHeight) * nNetworkWeight / nWeight) : 0;
